pyproject.toml
[project]
name = "litres-autotest-project"
version = "0.1.0"
description = ""
authors = [
    {name = "ilyashe",email = "ilya.shebanoff@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "selene (==2.0.0rc9)",
    "allure-pytest (>=2.14.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "appium-python-client (>=5.0.0,<6.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
